+++ apps/next/pages/edit-profile/index.tsx
@@ -1,0 +1,11 @@
+import { AuthGuard } from "app/utils/authGuard";
+import EditProfileScreen from "app/features/edit-profile";
+import React from "react";
+
+export default function EditProfilePage() {
+  return (
+    <AuthGuard>
+      <EditProfileScreen />
+    </AuthGuard>
+  );
+}
