+++ packages/app/provider/navigation/index.tsx
@@ -1,13 +1,9 @@
 import { NavigationContainer } from "@react-navigation/native";
 import * as Linking from "expo-linking";
-import { useMemo } from "react";
+import { ReactNode, useMemo } from "react";
 import { useAuthStatus } from "app/hooks/useAuthStatus";
 
-export function NavigationProvider({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
+export function NavigationProvider({ children }: { children: ReactNode }) {
   useAuthStatus();
 
   return (
@@ -24,6 +20,7 @@ export function NavigationProvider({
               home: "home",
               "sign-in": "sign-in",
               "sign-up": "sign-up",
+              "edit-profile": "edit-profile",
               dashboard: {
                 screens: {
                   search: "dashboard/search",
